+++ packages/vant-cli/src/compiler/compile-sfc.ts
@@ -87,7 +87,7 @@ export async function compileSfc(filePath: string): Promise<any> {
   const descriptor = parseSfc(filePath);
   const { template, styles } = descriptor;
 
-  const hasScoped = styles.some(s => s.scoped);
+  const hasScoped = styles.some((s) => s.scoped);
   const scopeId = hasScoped ? `data-v-${hash(source)}` : '';
 
   // compile js part
@@ -107,9 +107,7 @@ export async function compileSfc(filePath: string): Promise<any> {
         }
 
         writeFileSync(jsFilePath, script);
-        compileJs(jsFilePath)
-          .then(resolve)
-          .catch(reject);
+        compileJs(jsFilePath).then(resolve).catch(reject);
       })
     );
   }