+++ packages/vant-markdown-vetur/src/parser.ts
@@ -27,7 +27,7 @@ function readLine(input: string) {
 function splitTableLine(line: string) {
   line = line.replace('\\|', 'JOIN');
 
-  const items = line.split('|').map(item => item.trim().replace('JOIN', '|'));
+  const items = line.split('|').map((item) => item.trim().replace('JOIN', '|'));
 
   // remove pipe character on both sides
   items.pop();