+++ src/action-sheet/index.tsx
@@ -75,12 +75,6 @@ function ActionSheet(
     }
   }
 
-  function Content() {
-    if (slots.default) {
-      return <div class={bem('content')}>{slots.default()}</div>;
-    }
-  }
-
   function Option(item: ActionSheetItem, index: number) {
     const { disabled, loading, callback } = item;
 
@@ -161,8 +155,10 @@ function ActionSheet(
     >
       {Header()}
       {Description()}
-      {props.actions && props.actions.map(Option)}
-      {Content()}
+      <div class={bem('content')}>
+        {props.actions && props.actions.map(Option)}
+        {slots.default?.()}
+      </div>
       {CancelText()}
     </Popup>
   );