+++ src/calendar/demo/index.vue
@@ -121,8 +121,6 @@
 </template>
 
 <script>
-import { GREEN } from '../../utils/constant';
-
 export default {
   i18n: {
     'zh-CN': {
@@ -235,7 +233,7 @@ export default {
           this.showConfirm = false;
           break;
         case 'customColor':
-          this.color = GREEN;
+          this.color = '#1989fa';
           break;
         case 'customConfirm':
           this.confirmText = this.t('confirmText');