+++ src/calendar/index.js
@@ -112,7 +112,6 @@ export default createComponent({
     return {
       subtitle: '',
       currentDate: this.getInitialDate(),
-      realRowHeight: 0,
     };
   },
 
@@ -266,28 +265,34 @@ export default createComponent({
 
       let height = 0;
       let currentMonth;
-      let visibleIndex;
+      const visibleRange = [-1, -1];
 
       for (let i = 0; i < months.length; i++) {
         const visible = height <= bottom && height + heights[i] >= top;
 
-        if (visible && !currentMonth) {
-          visibleIndex = i;
-          currentMonth = months[i];
-        }
+        if (visible) {
+          visibleRange[1] = i;
 
-        if (!months[i].visible && visible) {
-          this.$emit('month-show', {
-            date: months[i].date,
-            title: months[i].title,
-          });
+          if (!currentMonth) {
+            currentMonth = months[i];
+            visibleRange[0] = i;
+          }
+
+          if (!months[i].showed) {
+            months[i].showed = true;
+            this.$emit('month-show', {
+              date: months[i].date,
+              title: months[i].title,
+            });
+          }
         }
 
         height += heights[i];
       }
 
       months.forEach((month, index) => {
-        month.visible = index >= visibleIndex - 1 && index <= visibleIndex + 1;
+        month.visible =
+          index >= visibleRange[0] - 1 && index <= visibleRange[1] + 1;
       });
 
       /* istanbul ignore else */
@@ -399,10 +404,6 @@ export default createComponent({
       this.$emit('confirm', copyDates(this.currentDate));
     },
 
-    onUpdateHeight(height) {
-      this.realRowHeight = height;
-    },
-
     genMonth(date, index) {
       const showMonthTitle = index !== 0 || !this.showSubtitle;
       return (
@@ -421,11 +422,9 @@ export default createComponent({
           currentDate={this.currentDate}
           showSubtitle={this.showSubtitle}
           allowSameDay={this.allowSameDay}
-          realRowHeight={this.realRowHeight}
           showMonthTitle={showMonthTitle}
           firstDayOfWeek={this.dayOffset}
           onClick={this.onClickDay}
-          onUpdate-height={this.onUpdateHeight}
         />
       );
     },