+++ src/picker/PickerColumn.js
@@ -166,7 +166,7 @@ export default createComponent({
       this.setIndex(index, true);
 
       // compatible with desktop scenario
-      // use setTimeout to skip the click event triggered after touchstart
+      // use setTimeout to skip the click event Emitted after touchstart
       setTimeout(() => {
         this.moving = false;
       }, 0);
@@ -305,7 +305,7 @@ export default createComponent({
 
         return (
           <li {...data}>
-            <div {...childData} />
+            {this.slots('option', option) || <div {...childData} />}
           </li>
         );
       });