+++ src/sku/Sku.js
@@ -73,6 +73,10 @@ export default createComponent({
       type: Boolean,
       default: true,
     },
+    disableSoldoutSku: {
+      type: Boolean,
+      default: true,
+    },
     customStepperConfig: {
       type: Object,
       default: () => ({}),
@@ -655,6 +659,7 @@ export default createComponent({
       stepperTitle,
       selectedSkuComb,
       showHeaderImage,
+      disableSoldoutSku,
     } = this;
 
     const slotsProps = {
@@ -720,6 +725,7 @@ export default createComponent({
                   skuKeyStr={skuTreeItem.k_s}
                   selectedSku={selectedSku}
                   skuEventBus={skuEventBus}
+                  disableSoldoutSku={disableSoldoutSku}
                   largeImageMode={skuTreeItem.largeImageMode}
                 />
               ))}