+++ src/switch-cell/index.tsx
@@ -30,6 +30,12 @@ function SwitchCell(
   slots: DefaultSlots,
   ctx: RenderContext<SwitchCellProps>
 ) {
+  if (process.env.NODE_ENV === 'development') {
+    console.warn(
+      '[Vant] "SwitchCell" component is deprecated, see: https://youzan.github.io/vant/#/zh-CN/switch-cell.'
+    );
+  }
+
   return (
     <Cell
       center